% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Make chart with plots of accidents locations in some state for a year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer}

\item{year}{in four digits}
}
\value{
plot chart
}
\description{
If state number doesn't exist error will be thrown.
If there's no accidents in state for year message will be returned
Otherwise map of state will accident locations will be created.
All locations has to be within boundaries.
}
\examples{
\dontrun{
fars_map_state(1, 2013)
}
}
