% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Add year variable to data, and return month, year list}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years}
}
\value{
list
}
\description{
It will add year variable for all requested years data, and return a list.
In case that there's no data file for a year, warrning will be returned.
Data files has to be in the same directory as this script.
}
\examples{
fars_read_years(2013)
fars_read_years(c('2013', 2014))
fars_read_years(c('2013', 2018, 2019))
}
