% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Get count of records for each year by month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years}
}
\value{
data.frame
}
\description{
For each requested year it will calculate count of records. Results for all
years and months will be returned as data.frame
In case that data from one of years doesn't exist warning will be returned
for that year only.
}
