% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generates file name based on year parameter}
\usage{
make_filename(year)
}
\arguments{
\item{year}{in four digits}
}
\value{
string
}
\description{
If year parameter can't be converted to integer error will be thrown.
}
\examples{
\dontrun{
make_filename(2018)
make_filename('2018')
}
}
